tests: pytest: added session scoped autoused fixture to prepare env
This reduces the number of tests failing with pytest from 53 to 11.
